Initialise subflows structure at SYN packet
The MPTCP RFC states that the fallback to tcp
should happen only when a data ack with no dss
is encountered, and not at the 3rd ack.

It means that an 3WH of:
- syn with mp_capable
- syn/ack with mp_capable
- ack without mp_capable

should still result in mptcp data transfer if
the dss is correct in subsequent packets.

This patch makes this scenario testable with
packetdrill_mptcp.